src/cloud_run/mimitalk_dev/requirements.txt
functions-framework==3.*
Flask
google-cloud-speech
google-cloud-texttospeech
google-cloud-firestore
google-api-core
google-cloud-aiplatform[agents]